[Fix] CheckPoint Position
- 체크포인트 위치의 y이값이 이상하게 생성되는 버그 수정
- Rewind하고 캐릭터의 y값이 이상하게 되는 버그 수정

+++ BeatMasterProject/Assets/02. Scripts/LevelGame/ObjectGenerator.cs
@@ -131,7 +131,7 @@ public class ObjectGenerator : MonoBehaviour
     /// </summary>
     /// <param name="xPos"></param>
     /// <param name="yPos"></param>
-    public void RecordCheckPoint(int xPos, int yPos)
+    public void RecordCheckPoint(float xPos, float yPos)
     {
         checkPointPos.Add(new Vector3(xPos, yPos, 0)); // Record position
     }
